% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_filter.R
\name{apply_filter}
\alias{apply_filter}
\title{Apply filter logic to user data}
\usage{
apply_filter(user, filter_params)
}
\arguments{
\item{user}{user table from database}

\item{filter_params}{a string specifying the filter to apply}
}
\value{
user data appended with incl column.
}
\description{
Apply filter logic to user data
}
