% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_audience.R
\name{create_audience}
\alias{create_audience}
\title{Create records in audience table and audience_filter}
\usage{
create_audience(name, audience_filter, dataset = Sys.getenv("bq_dataSet"))
}
\arguments{
\item{name}{Audience name}

\item{audience_filter}{Audience filters}

\item{dataset}{BigQuery dataset}
}
\value{
audience_id
}
\description{
Create records in audience table and audience_filter
}
