% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create record in experiment table}
\usage{
create_experiment(
  input,
  rv,
  session,
  audience_id,
  treatment_id,
  dataset = Sys.getenv("bq_dataSet")
)
}
\arguments{
\item{audience_id}{audience to target with experiment}

\item{dataset}{BigQuery dataset}

\item{name}{experiment name}

\item{experiment_treatment}{treatments in the experiment, with weightings}

\item{start_datetime}{when the experiment will start}

\item{end_datetime}{when the experiment will end}
}
\value{
experiment_id
}
\description{
Create record in experiment table
}
