% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_treatment.R
\name{create_treatment}
\alias{create_treatment}
\title{Create record in treatment table}
\usage{
create_treatment(name, dataset = Sys.getenv("bq_dataSet"))
}
\arguments{
\item{name}{name for new treatment}

\item{dataset}{BigQuery project ID}
}
\value{
treatment_id
}
\description{
Create record in treatment table
}
