% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_user_metadata.R
\name{create_user_metadata}
\alias{create_user_metadata}
\title{Create user metadata}
\usage{
create_user_metadata(user = NULL)
}
\arguments{
\item{user}{\itemize{
\item user_data (from get_user)
}}
}
\value{
list of user metrics, variables, and variable values
}
\description{
Create user metadata
}
