% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_experiment_treatment.R
\name{format_experiment_treatment}
\alias{format_experiment_treatment}
\title{Format experiment_treatment}
\usage{
format_experiment_treatment(experiment_treatment, treatment)
}
\arguments{
\item{experiment_treatment}{experiment_treatment table from database}

\item{treatment}{treatment table from database}
}
\value{
experiment_treatment formatted to be more readable
}
\description{
Format experiment_treatment
}
