% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment_summary.R
\name{get_experiment_summary}
\alias{get_experiment_summary}
\title{Get summary of experiment}
\usage{
get_experiment_summary(
  experiment_name,
  experiment_audience,
  experiment,
  treatment
)
}
\arguments{
\item{experiment_name}{experiment name}

\item{experiment_audience}{experiment audience from database}

\item{experiment}{experiment from database}
}
\value{
String with time live (or until start) and users assigned each treatment
}
\description{
Get summary of experiment
}
