% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Pull table from database}
\usage{
get_table(table_name, limit = NULL, dataset = Sys.getenv("bq_dataSet"))
}
\arguments{
\item{table_name}{table to pull}

\item{limit}{maximum number of rows to pull}

\item{dataset}{BigQuery project ID}
}
\value{
table
}
\description{
Pull table from database
}
