% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_impact.R
\name{get_user_impact_data}
\alias{get_user_impact_data}
\title{Pulls user-level data to analyze test results}
\usage{
get_user_impact_data(
  dataset = Sys.getenv("bq_dataSet"),
  experiment_id,
  impact_variable,
  pre_period_days = 14
)
}
\arguments{
\item{dataset}{Big Query dataset}

\item{impact_variable}{variable to be measured}

\item{pre_period_days}{how far into the past to pull pre-period data}

\item{experiment_name}{experiment to analyze}
}
\value{
user_impact_data
}
\description{
Pulls user-level data to analyze test results
}
