% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_experiment.R
\name{input_to_experiment_record}
\alias{input_to_experiment_record}
\title{Maps shiny inputs to format in experiment database table}
\usage{
input_to_experiment_record(input)
}
\arguments{
\item{input}{shiny input object}

\item{audience}{local version of audience table}
}
\value{
experiment parameters in a format consistent with database
}
\description{
Maps shiny inputs to format in experiment database table
}
