% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingModal.R
\name{loadingModal}
\alias{loadingModal}
\title{Utility function for easy modal dialog}
\usage{
loadingModal(text = "Loading...")
}
\arguments{
\item{text}{Test to display in the modal}
}
\value{
simple modal with displaying text
}
\description{
Utility function for easy modal dialog
}
