% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_impact.R
\name{measure_user_impact}
\alias{measure_user_impact}
\title{Analyzes output of get_user_impact_data}
\usage{
measure_user_impact(user_impact_data)
}
\arguments{
\item{user_impact_data}{Output of get_user_impact_data}
}
\value{
text description of test results
}
\description{
Analyzes output of get_user_impact_data
}
