% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_impact.R
\name{plot_timeseries_impact}
\alias{plot_timeseries_impact}
\title{Plots timeseries impact}
\usage{
plot_timeseries_impact(
  timeseries_impact_data,
  experiment_start,
  colors = rctr_colors()
)
}
\arguments{
\item{timeseries_impact_data}{output of get_timeseries_impact_data}

\item{colors}{line colors for plot}

\item{impact_variable}{impact variable name (for plot_)}
}
\value{

}
\description{
Plots timeseries impact
}
