% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_data}
\alias{pull_data}
\title{Pull data from database}
\usage{
pull_data(sql, projectID = Sys.getenv("bq_projectID"))
}
\arguments{
\item{sql}{sql statement to run}

\item{projectID}{BigQuery project ID, by default will use Sys.getenv("bq_projectID")}
}
\value{
data.table populated with query results
}
\description{
Evaluate sql statement and return result as a data.table. Ensure that you
have authorized with BigQuery e.g., with bigrquery::bq_auth().
}
