% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push_data.R
\name{push_data}
\alias{push_data}
\title{push data to database}
\usage{
push_data(
  tableId,
  upload_data,
  dataSet = Sys.getenv("bq_dataSet"),
  projectID = Sys.getenv("bq_projectID"),
  writeDisposition = "WRITE_APPEND",
  schema = NULL
)
}
\arguments{
\item{tableId}{Name of table in GCP}

\item{upload_data}{data.table to upload}

\item{dataSet}{name of BigQuery data set}

\item{projectID}{BigQuery project ID}

\item{writeDisposition}{see ?bigQueryR::bqr_upload_data}

\item{schema}{see ?bigQueryR::bqr_upload_data}
}
\description{
Wrapper for bigQueryR::bqr_upload_data that uses BigQuery environment
variables by default.
}
\examples{
push_data("db_table", local_data_table)
}
