% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_experiment_summary.R
\name{get_experiment_summary}
\alias{get_experiment_summary}
\title{Get summary of experiment}
\usage{
get_experiment_summary(experiment_name)
}
\arguments{
\item{experiment_name}{}
}
\value{
String with time live (or until start) and users assigned each treatment
}
\description{
Get summary of experiment
}
