% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_impact.R
\name{get_timeseries_impact_data}
\alias{get_timeseries_impact_data}
\title{Pulls timeseries data for analysis and plotting}
\usage{
get_timeseries_impact_data(
  dataset = Sys.getenv("bq_dataSet"),
  experiment_id,
  impact_variable,
  pre_period_days = 14
)
}
\arguments{
\item{dataset}{BigQuery dataset}

\item{impact_variable}{variable to be measured}

\item{pre_period_days}{how far into the past to pull pre-period data}

\item{experiment_name}{experiment to analyze}
}
\value{
timeseries data
}
\description{
Pulls timeseries data for analysis and plotting
}
