% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries_impact.R
\name{plot_timeseries_impact}
\alias{plot_timeseries_impact}
\title{Plots timeseries impact}
\usage{
plot_timeseries_impact(
  timeseries_impact_data,
  experiment_name,
  experiment_start,
  impact_variable,
  colors = c("#F18E7E", "#69AEDB", "grey10")
)
}
\arguments{
\item{timeseries_impact_data}{output of get_timeseries_impact_data}

\item{experiment_name}{experiment name (for plot)}

\item{impact_variable}{impact variable name (for plot_)}

\item{colors}{line colors for plot}
}
\value{

}
\description{
Plots timeseries impact
}
